exploratory-requirements.txt
numpy~=1.21.2
pandas~=1.3.3
pillow~=8.3.1
tabulate~=0.8.9
click~=8.0.1
opencv-python~=3.4.8.29
setuptools~=58.0.4